requirements.txt
pyTelegramBotAPI==4.12.0
python-decouple==3.8
python_dateutil==2.8.2
six==1.16.0